Rename Grid/qcd/utils/<Group>_impl.h -> Grid/qcd/utils/<Group>.h
